Implement 'getKnowledgeBaseDetail'. - v1.0

+++ knowledge-manager-core/src/main/java/ifl/agentbreaker/knowledgemanager/domain/dtos/responses/DocumentDetailResponse.java
@@ -55,7 +55,7 @@ public class DocumentDetailResponse
     /**
      * Chunk数量
      */
-    private Integer chunkCount;
+    private long chunkCount;
 
     /**
      * 上传时间
