+++ knowledge-manager-core/src/main/java/ifl/agentbreaker/knowledgemanager/domain/dtos/responses/DocumentResponse.java
@@ -50,7 +50,7 @@ public class DocumentResponse
     /**
      * chunk数量
      */
-    private Integer chunkCount;
+    private long chunkCount;
 
     /**
      * 上传时间
